Added descriptions for Crit Bonus and Discount researchers
The current event features a Crit Bonus researcher that the tracker doesn't know how to describe.  I have fixed that and also included an implementation for Reduction.  Technically, this can't be seen anywhere on the tracker currently (the only potential place would be if it was motherland scripted mission reward, and they never are), but could be useful in the future.