Fixed missing icons in calculator popup.
My commit from last week that reorganized the ./img/event/ folder did not also change the url's in the calculator, so they were broken links.  This commit fixes that.
I am committing directly to master in-browser because I consider this a medium-high priority issue and I am currently out of town.