Initial researcher commit - UI
Separated the the UI for calc's input into a Generator counts sub-tab, and a Researcher levels sub-tab.  You will still have to type in counts, but derived values like power you now set the appropriate researchers' levels, which is much more intuitive.
The calc is NOT FUNCTIONAL yet, this commit is merely a checkpoint where the UI is usable.
Added the multi-industry icon to each theme, added a propaganda boost pseudo-researcher image, and added all researchers from motherland and cowboy in for testing.  The rest will come later.