Updated santa + schedule (mostly)
Updated santa-bal-10 and the Schedule with the latest version.  Not many user-facing changes, but lots of internal stuff like specops.
This includes some changes to missions.js to handle scheduling better, but is still uncomplete and will start showing incorrect events after the Santa event is done unless the code is completed by then.