Improved blank thousand separator
Previously, I just had a default value for the thousand separator if it was blank, but this could potentially lead to the thousand and decimal separator being the same (if thousand is "" and decimal is ",").  Now a blank thousand separator ensures that it will not be equal to the decimal seperator when it gets a default value.