Added fake Santa entry
HH is taking a bit too long to update the internal schedule for my comfort, so I'm breaking my typical policy of not editing data to include it.  It both has a notice explaining it to users and it's confined to a specific block of code separate from the actual json to be easily removed.