Added Max Sim Time input
Added a Max Sim Time input option to the normal production sim calculator, which configures the number of real life seconds the simulation should run before terminating.  This replaces the previously-removed Max Days input that ended at a specific in-simulation time, and the recently-added static 10-second real-time limit, which was too long for typical failure cases.